Move implementation to patch file
